classifier/requirements.txt
fastapi==0.103.2
opencv-python==4.8.1.78
Pillow==9.4.0
streamlit==1.27.2
matplotlib==3.7.3
st-annotated-text
mlflow==2.7.1
protobuf==3.20